feat: unifica card de boas-vindas com benefícios de apoiador
